create new file to import wizard_account_product_fiscal_classification, account_product_fiscal_classification_template
