+++ l10n_br_account_product/models/__init__.py
@@ -19,7 +19,9 @@ from . import l10n_br_account_product_cest
 from . import res_company
 from . import l10n_br_tax_definition_company_product
 from . import l10n_br_account
+from . import account_product_fiscal_classification_template
 from . import account_product_fiscal_classification
+from . import wizard_account_product_fiscal_classification
 from . import account_invoice
 from . import account_invoice_line
 from . import account_invoice_tax
